requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
openai==1.3.0
eventlet==0.33.3
flask-socketio==5.3.5
flask-cors==4.0.0
flask-mail==0.9.1
python-dotenv==1.0.0
Pillow==10.0.0
httpx==0.24.1
markupsafe==2.1.3
pathlib==1.0.1
gunicorn==21.2.0
ffmpeg-python==0.2.0
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2